Corrected 'is_adjacent' function within nbody-bins
First off, there's a duplicate comparison for +1. Secondly,
the intent seems to be a symmetric comparison for both bins.
The last 4 lines were duplicates and the 4 lines before that
did not have positive counterparts. Those lines and the
comment were updated to be inline with the commutative
nature of this function.