lectures/L23-slides: Add slide on F64 vs F32 performance on GPUs
Many GPUs don't support native FP64 computation, rather emulating it
using FP32. As such, performance of GPU based applications can typically
be improved (typically 32x) by trading the accuracy of a 64 bit float to
improve performance using hardware supported 32 bit floats.

This commit adds a slide outlining this difference and optional tradeoff
as well as adding a small table of FLOPS values for common GPUs. While
I'm not sure if this is in-scope for this course, given the discussion
of performance, it can be a noteworthy factor.

Signed-off-by: Tony Tascioglu <tony@tonytascioglu.com>